Integrate work from forks by @MsKrypt1 and @fgsfdsfgs
This commit integrates all of the additions that have been done since the previous commit on this repository by users @MsKrypt1 and @fgsfdsfgs on their separate forks. As it was not possible to get more contact info from them in order to solicit pull requests, a decision was made to manually add their contents into the upstream repository. The many forks were creating a rabbit hole in regards to properly following the decompilation status, and should this measure not be agreed upon by the aforementioned forkers, please leave a comment here to start discussing matters.

Co-Authored-By: MsKrypt1 <19950257+mskrypt1@users.noreply.github.com>
Co-Authored-By: fgsfds <2997482+fgsfdsfgs@users.noreply.github.com>